requirements.txt
# Simple requirements - let pip resolve versions
fastapi
uvicorn[standard]
httpx
python-dotenv
pytest